requirements.txt
absl-py==1.1.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiostream==0.4.4
anyio==3.6.1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
bert-serving-client==1.10.0
bert-serving-server==1.10.0
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
clip-client==0.4.20
clip-server==0.4.20
commonmark==0.9.1
cryptography==37.0.3
cycler==0.11.0
docarray==0.13.29
docker==5.0.3
elastic-transport==8.1.2
elasticsearch==6.8.2
fastapi==0.78.0
filelock==3.7.1
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
ftfy==6.1.1
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.47.0
grpcio-health-checking==1.47.0
grpcio-reflection==1.47.0
grpcio-tools==1.47.0
h11==0.13.0
h5py==3.7.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.4.0
jina==3.6.9
joblib==1.1.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
lz4==4.0.1
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
nltk==3.7
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==9.2.0
prometheus-client==0.14.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
rake-nltk==1.0.6
regex==2022.6.2
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.4.4
rsa==4.8
sacremoses==0.0.53
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
wcwidth==0.2.5
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
